refactor: reorganize files
- move all code into lib folder
- move unit tests next to lib code
- rename files in kebab-case
- sort some requires
- remove utilities folder since there was only one file
