+++ lib/bucket-utils.js
@@ -56,17 +56,18 @@ function emptyBucket(aws, bucketName, keyPrefix) {
     if (!contents[0]) {
       return Promise.resolve();
     } else {
-
       if (keyPrefix) {
         testPrefix = true;
         prefixRegexp = new RegExp('^' + keyPrefix);
       }
-      const objects = contents.map(function(content) {
-        return {Key: content.Key};
-      }).filter(content => !testPrefix || prefixRegexp.test(content.Key));
+      const objects = contents
+        .map(function (content) {
+          return { Key: content.Key };
+        })
+        .filter(content => !testPrefix || prefixRegexp.test(content.Key));
 
       if (objects.length === 0) {
-        return Promise.resolve()
+        return Promise.resolve();
       }
 
       const params = {
