+++ lib/configure.js
@@ -122,7 +122,7 @@ function configurePolicyForBucket(aws, bucketName, customPolicy) {
 function configureCorsForBucket(aws, bucketName) {
   const params = {
     Bucket: bucketName,
-    CORSConfiguration: require('./resources/CORSPolicy')
+    CORSConfiguration: require('./resources/cors-policy.json')
   };
 
   return aws.request('S3', 'putBucketCors', params);
