+++ lib/get-file-list.js
