+++ lib/plugin.js
@@ -1,15 +1,14 @@
 'use strict';
 
-const path = require('path');
-const fs = require('fs');
-
 const confirm = require('@serverless/utils/inquirer/confirm');
+const fs = require('fs');
+const path = require('path');
 
-const bucketUtils = require('./lib/bucketUtils');
-const configure = require('./lib/configure');
-const regionUrls = require('./lib/resources/awsRegionUrls');
-const uploadDirectory = require('./lib/upload');
-const validateClient = require('./lib/validate');
+const bucketUtils = require('./bucket-utils');
+const configure = require('./configure');
+const regionUrls = require('./resources/aws-region-urls.json');
+const uploadDirectory = require('./upload');
+const validateClient = require('./validate');
 
 class Client {
   constructor(serverless, cliOptions, { log }) {
