+++ lib/plugin.test.js
@@ -5,7 +5,7 @@ const inquirer = require('@serverless/utils/inquirer');
 const chai = require('chai');
 const sinon = require('sinon');
 
-const runServerless = require('../run-serverless');
+const runServerless = require('../test/run-serverless');
 
 chai.use(require('chai-as-promised'));
 chai.use(require('sinon-chai'));
