+++ lib/resources/aws-region-urls.json
