+++ lib/resources/cors-policy.json
