+++ lib/upload.js
@@ -1,14 +1,11 @@
 'use strict';
 
 const fs = require('fs');
+const mime = require('mime');
+const minimatch = require('minimatch');
 const path = require('path');
 
-const mime = require('mime');
-
-const getFileList = require('./utilities/getFileList');
-const regionUrls = require('./resources/awsRegionUrls');
-
-const minimatch = require('minimatch');
+const getFileList = require('./get-file-list');
 
 /**
  * Uploads client files to an S3 bucket
