+++ lib/validate.js
@@ -1,9 +1,8 @@
 'use strict';
 
 const fs = require('fs');
-const path = require('path');
-
 const is = require('is_js');
+const path = require('path');
 
 /**
  * Validates the configuration parameters that will be used for deployment
